Allow users.body & additional fields to be null (#481)
This changes multiple columns to allow nulls and default to `null`. 

* `users.body`
* `users.preferences`
* `users.longitude` and `users.latitude`
* `settings.locale`
* `groups.cover`

Fixes #480

Because this is the first time this app is modifying a column in a
migration, we need a new dependency:
> Before modifying a column, you must install the doctrine/dbal package
using the Composer package manager. The Doctrine DBAL library is used to
determine the current state of the column and to create the SQL queries
needed to make the requested changes to your column

from Laravel docs: https://laravel.com/docs/8.x/migrations

While `doctrine/dbal` is currently on 4.x, that is not compatible with
Laravel 8.x, so the version is locked to 3.x and this is noted in the
commit message.

This PR also rebuilds Composer for the first time since #479 & #476
because it is required for the new migration to build.