Manual coding standards & static analysis fixes (#650)
- Fix the rest of the PSR-12 coding standards issues.
- Fix test namespace & ref issues.
- Fix a few basic static analysis errors.

Branched from #649 (merge that first!)

For the record, I'm happy to use snake_case for test methods per the
PHPUnit convention, but today there is a mix and snake_case is in the
minority, so I'm setting them to PSR-12 for the moment.

I think it's relevant to mention that I found #652, #653, and #654 by
working on standards & static analysis. I am sure we'll find more as we
increase the level of PHPStan.